ddply na.omit only if na group

+++ src/main/java/water/exec/ASTRApply.java
@@ -264,27 +264,36 @@ class ASTddply extends ASTOp {
     Frame ff = new Frame(names,vres);
 
     // Cleanup pass: Drop NAs (groups with no data and NA groups, basically does na.omit: drop rows with NA)
-    Frame res = new MRTask2() {
-      @Override public void map(Chunk[] cs, NewChunk[] nc) {
-        int rows = cs[0].len();
-        int cols = cs.length;
-        boolean[] NACols = new boolean[cols];
-        ArrayList<Integer> xrows = new ArrayList<Integer>();
-        for (int i = 0; i < cols; ++i) NACols[i] = (cs[i]._vec.naCnt() != 0);
-        for (int r = 0; r < rows; ++r)
-          for (int c = 0; c < cols; ++c)
-            if (NACols[c])
-              if (cs[c].isNA0(r)) { xrows.add(r); break;}
-        for (int r = 0; r < rows; ++r) {
-          if (xrows.contains(r)) continue;
-          for (int c = 0; c < cols; ++c) {
-            if (cs[c]._vec.isEnum()) nc[c].addEnum((int) cs[c].at80(r));
-            else nc[c].addNum(cs[c].at0(r));
+    boolean anyNA = false;
+    Frame res = ff;
+    for (Vec v : ff.vecs()) if (v.naCnt() != 0) { anyNA = true; break; } // stop on first vec with naCnt != 0
+    if (anyNA) {
+      res = new MRTask2() {
+        @Override
+        public void map(Chunk[] cs, NewChunk[] nc) {
+          int rows = cs[0]._len;
+          int cols = cs.length;
+          boolean[] NACols = new boolean[cols];
+          ArrayList<Integer> xrows = new ArrayList<Integer>();
+          for (int i = 0; i < cols; ++i) NACols[i] = (cs[i]._vec.naCnt() != 0);
+          for (int r = 0; r < rows; ++r)
+            for (int c = 0; c < cols; ++c)
+              if (NACols[c])
+                if (cs[c].isNA0(r)) {
+                  xrows.add(r);
+                  break;
+                }
+          for (int r = 0; r < rows; ++r) {
+            if (xrows.contains(r)) continue;
+            for (int c = 0; c < cols; ++c) {
+              if (cs[c]._vec.isEnum()) nc[c].addEnum((int) cs[c].at80(r));
+              else nc[c].addNum(cs[c].at0(r));
+            }
           }
         }
-      }
-    }.doAll(ff.numCols(), ff).outputFrame(null, ff.names(), ff.domains());
-    ff.delete();
+      }.doAll(ff.numCols(), ff).outputFrame(null, ff.names(), ff.domains());
+      ff.delete();
+    }
     // Delete the group row vecs
     UKV.remove(envkey);
     env.poppush(4,res,null);
